HHH-17406 Change EntityPersister mutation api to expose coordinators
