+++ hibernate-core/src/main/java/org/hibernate/id/insert/UniqueKeySelectingDelegate.java
@@ -77,6 +77,7 @@ public class UniqueKeySelectingDelegate extends AbstractSelectingDelegate {
 		}
 	}
 
+	@Override
 	protected String getSelectSQL() {
 		return selectString;
 	}
